[aat] Move get_class to mod
