+++ src/hb/aat/layout_morx_table.rs
@@ -1,6 +1,6 @@
+use super::get_class;
 use super::layout::*;
 use super::map::{AatMap, AatMapBuilder, RangeFlags};
-use super::ClassCache;
 use crate::hb::aat::layout_common::AatApplyContext;
 use crate::hb::ot_layout::MAX_CONTEXT_LENGTH;
 use crate::hb::{hb_font_t, GlyphInfo};
@@ -9,7 +9,7 @@ use read_fonts::tables::aat::{ExtendedStateTable, NoPayload, StateEntry};
 use read_fonts::tables::morx::{
     ContextualEntryData, ContextualSubtable, InsertionEntryData, LigatureSubtable, SubtableKind,
 };
-use read_fonts::types::{BigEndian, FixedSize, GlyphId, GlyphId16};
+use read_fonts::types::{BigEndian, FixedSize, GlyphId16};
 
 // TODO: [morx] Blocklist dysfunctional morx table of AALMAGHRIBI.ttf font
 // HarfBuzz commit 1e629c35113e2460fd4a77b4fa9ae3ff6ec876ba
@@ -175,25 +175,6 @@ trait DriverContext<T> {
 }
 
 const START_OF_TEXT: u16 = 0;
-
-fn get_class<T: bytemuck::AnyBitPattern + FixedSize + core::fmt::Debug>(
-    machine: &ExtendedStateTable<'_, T>,
-    glyph_id: GlyphId,
-    cache: Option<&ClassCache>,
-) -> u16 {
-    if let Some(cache) = cache {
-        if let Some(klass) = cache.get(glyph_id.to_u32()) {
-            return klass as u16;
-        }
-    }
-    let klass = machine
-        .class(glyph_id)
-        .unwrap_or(read_fonts::tables::aat::class::OUT_OF_BOUNDS as u16);
-    if let Some(cache) = cache {
-        cache.set(glyph_id.to_u32(), klass as u32);
-    }
-    klass
-}
 
 fn drive<T: bytemuck::AnyBitPattern + FixedSize + core::fmt::Debug>(
     machine: &ExtendedStateTable<'_, T>,
